Fix unwrap to handle null bytes
When unwrapping with an incorrect key, it is likely the (incorrectly)
decrypted message (effectively random binary) includes a null byte.
When trying to capture this with command substitution (e.g.,
`foo="$(...)"`), bash will ignore the null byte. The result is:
- The wrong value stored
- An uncapturable error condition
- An unsuppressable error message printed

This commit avoids the issue by passing the decrypted value through a
pipe to a test that more acurately checks for a valid result.